DriveService and SheetsService classes made inner classes of helpers
